# packages/shared-types/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/work-orders": {
      "post": {
        "operationId": "WorkOrdersController_create_v1",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Workshop creates a work order (customer by phone, vehicle by id/VIN/plate, optional items)",
        "tags": [
          "work-orders"
        ]
      },
      "get": {
        "operationId": "WorkOrdersController_list_v1",
        "parameters": [
          {
            "name": "org_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List: ?org_id= for the workshop, otherwise my (customer/fleet) orders; ?status=a,b",
        "tags": [
          "work-orders"
        ]
      }
    },
    "/v1/work-orders/{id}": {
      "get": {
        "operationId": "WorkOrdersController_get_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "work-orders"
        ]
      },
      "patch": {
        "operationId": "WorkOrdersController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "work-orders"
        ]
      }
    },
...
    "/v1/accident-reports/lookup": {
      "post": {
        "operationId": "AccidentsController_lookup_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Preview an accident report by its reference (منجز/تقدير) — nothing is stored",
        "tags": [
          "accident-reports"
        ]
      }
    },
    "/v1/accident-reports": {
      "post": {
        "operationId": "AccidentsController_link_v1",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Link the report to a work order: draft items, insurer amounts, Car Passport event",
        "tags": [
          "accident-reports"
        ]
      },
      "get": {
        "operationId": "AccidentsController_list_v1",
        "parameters": [
          {
            "name": "org_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vehicle_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "accident-reports"
        ]
      }
    },
    "/v1/accident-reports/{id}": {
      "get": {
        "operationId": "AccidentsController_get_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "accident-reports"
        ]
      }
    },
    "/v1/accident-reports/{id}/refresh": {
      "post": {
        "operationId": "AccidentsController_refresh_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Re-read the file from the provider (the assessment lands days later)",
        "tags": [
          "accident-reports"
        ]
      }
    },
    "/v1/accident-reports/{id}/submit-repair": {
      "post": {
        "operationId": "AccidentsController_submitRepair_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "FR-WO-10 — register the completed repair against the accident file",
        "tags": [
          "accident-reports"
        ]
      }
    },
    "/v1/work-orders/{id}/accident-report": {
      "get": {
        "operationId": "WorkOrderAccidentController_byWorkOrder_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The accident file behind this repair, if any",
        "tags": [
          "work-orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "Sinaaty API — صناعتي",
    "description": "",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}